requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
